refactor: move getArrivalTime method to SubwayPath - remove unnecessary TimePath class
