+++ src/main/java/nextstep/subway/maps/map/domain/TimePaths.java
@@ -1,27 +1,25 @@
-package nextstep.subway.maps.map.domain.path;
+package nextstep.subway.maps.map.domain;
 
 import java.time.LocalDateTime;
+import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
-import java.util.stream.Collectors;
 
-import nextstep.subway.maps.map.domain.SubwayPath;
+import com.google.common.collect.Lists;
 
 public class TimePaths {
 
-    private final List<TimePath> paths;
+    private final List<SubwayPath> paths;
 
     private TimePaths(List<SubwayPath> paths) {
-        this.paths = paths.stream()
-            .map(TimePath::new)
-            .collect(Collectors.toList());
+        this.paths = Collections.unmodifiableList(Lists.newArrayList(paths));
     }
 
     public static TimePaths of(List<SubwayPath> paths) {
         return new TimePaths(paths);
     }
 
-    public TimePath findFastestArrivalPath(LocalDateTime departTime) {
+    public SubwayPath findFastestArrivalPath(LocalDateTime departTime) {
         return this.paths.stream()
             .min(Comparator.comparing(timePath -> timePath.getArrivalTime(departTime)))
             .orElseThrow(() -> new IllegalArgumentException("no paths found by given request from user."));
